requirements.txt
streamlit
ultralytics
pdf2image
pytesseract
Pillow
numpy
tqdm
seaborn
matplotlib
psutil
pyyaml